refactor: 이미지 업로드 컴포넌트로 분리
