+++ src/app/funding/write/page.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import ImageUpload from "@/components/funding/ImageUpload";
 import RegisterConfirmModal from "@/components/funding/RegisterConfirmModal";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
@@ -16,25 +17,11 @@ const FundingPage = () => {
     address,
     content,
     setTitle,
-    setImageUrl,
     setPurchaseLink,
     setAddress,
     setContent,
   } = useFundingStore();
   const [isRegister, setIsRegister] = useState(false);
-
-  const handleImageChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (file) {
-      const reader = new FileReader();
-      reader.onloadend = () => {
-        if (typeof reader.result === "string") {
-          setImageUrl(reader.result);
-        }
-      };
-      reader.readAsDataURL(file);
-    }
-  };
 
   const handleSubmit = () => {
     if (!title || !imageUrl || !purchaseLink || !address || !content) {
@@ -64,44 +51,7 @@ const FundingPage = () => {
           value={title}
           onChange={(e) => setTitle(e.target.value)}
         />
-        <div className="flex items-center gap-8">
-          <span className="text-sm text-secondary opacity-80">사진 첨부</span>
-          {imageUrl ? (
-            <div className="w-45 h-30 overflow-hidden">
-              <Image
-                src={imageUrl}
-                alt="Uploaded Image"
-                width={180}
-                height={120}
-                objectFit="contain"
-              />
-            </div>
-          ) : (
-            <label
-              htmlFor="picture"
-              className="flex inline-flex items-center cursor-pointer"
-            >
-              <div className="flex w-45 h-30 bg-background rounded-2xl justify-center items-center">
-                <div className="flex w-13 h-13 bg-primary rounded-4xl justify-center items-center">
-                  <Image
-                    src="/assets/icons/union.svg"
-                    alt="Union"
-                    width={25}
-                    height={25}
-                  />
-                </div>
-              </div>
-              <Input
-                id="picture"
-                type="file"
-                variant="picture"
-                size="xs"
-                style={{ display: "none" }}
-                onChange={handleImageChange}
-              />
-            </label>
-          )}
-        </div>
+        <ImageUpload />
         <Input
           type="text"
           size="sm"
